TranscritorReelsParaGitHub_v2/requirements.txt
streamlit
yt-dlp
openai-whisper
moviepy
imageio-ffmpeg